Remove configtest dependency from collector, generate tests in builder
